More flexible layout
My attempt at using a little HTML5 / CSS magic to get a better layout.  Instead of using absolute positioning, I switched to relative and put the whole thing in a grid (table) with the buttons between the boxes rather than overlayed along the box borders but translucent.

It's not perfect, but it manages to work - but importantly, no more having to manage a bunch of absolutely positioned buttons - no more having to tweak top/left positions - no more fine tuning.

I also included the previous version for reference/historical purposes.  Or, in case it is preferred for some reason.